Table View Established
The table populates with a Multi Column List View of the scriptable objects once the JSON file is placed into the tool. The columns are all locked to the appropriate field types for the variables they reference, albeit more work is needed for the sprite field. Lists do not work within the list view.