Saving Update
Can now save changes made in the tool to the SOs and the JSON file. Sprites and objects still do not save to the JSON, and therefore do not save to the SO either, but other datatypes are fine. Renaming an object will create a new SO without removing the old one.